Split binary op checking
