+++ src/typechecker/resolve.rs
@@ -1,10 +1,8 @@
-//! Type resolution — AstType → Type, field lookups, binary op checking.
+//! Type resolution — AstType → Type, field lookups, and compatibility checks.
 #![allow(clippy::result_large_err)]
 
-use crate::ast::expressions::BinaryOp;
 use crate::ast::typed::Type;
 use crate::ast::{is_builtin_type_name, AstType};
-use crate::error::{Diagnostic, Span};
 
 use super::TypeChecker;
 
@@ -185,117 +183,5 @@ impl TypeChecker {
         // Numeric width/sign conversions require explicit casts. Literal
         // coercion is handled before this check at declaration sites.
         false
-    }
-
-    pub(crate) fn check_binary_op(
-        &self,
-        op: BinaryOp,
-        left: &Type,
-        right: &Type,
-        span: &Span,
-    ) -> Result<Type, Diagnostic> {
-        match op {
-            BinaryOp::Add | BinaryOp::Sub | BinaryOp::Mul | BinaryOp::Div | BinaryOp::Mod => {
-                // Arithmetic — both sides must be numeric (allow Unknown for error recovery)
-                if *left == Type::Unknown || *right == Type::Unknown {
-                    let known = if *left != Type::Unknown {
-                        left
-                    } else {
-                        right
-                    };
-                    return Ok(known.clone());
-                }
-                if !left.is_numeric() {
-                    return Err(Diagnostic::error(
-                        "E3010",
-                        format!("arithmetic on non-numeric type `{}`", left.display_name()),
-                        *span,
-                    ));
-                }
-                if !right.is_numeric() {
-                    return Err(Diagnostic::error(
-                        "E3010",
-                        format!("arithmetic on non-numeric type `{}`", right.display_name()),
-                        *span,
-                    ));
-                }
-                if left != right {
-                    return Err(Diagnostic::error(
-                        "E3013",
-                        format!(
-                            "arithmetic operands must have the same type, found `{}` and `{}`",
-                            left.display_name(),
-                            right.display_name()
-                        ),
-                        *span,
-                    ));
-                }
-                Ok(left.clone())
-            }
-            BinaryOp::Eq
-            | BinaryOp::NotEq
-            | BinaryOp::Lt
-            | BinaryOp::Gt
-            | BinaryOp::LtEq
-            | BinaryOp::GtEq => Ok(Type::Bool),
-            BinaryOp::And | BinaryOp::Or => {
-                if *left != Type::Bool && *left != Type::Unknown {
-                    return Err(Diagnostic::error(
-                        "E3011",
-                        format!(
-                            "logical operator requires `bool`, found `{}`",
-                            left.display_name()
-                        ),
-                        *span,
-                    ));
-                }
-                if *right != Type::Bool && *right != Type::Unknown {
-                    return Err(Diagnostic::error(
-                        "E3011",
-                        format!(
-                            "logical operator requires `bool`, found `{}`",
-                            right.display_name()
-                        ),
-                        *span,
-                    ));
-                }
-                Ok(Type::Bool)
-            }
-            BinaryOp::BitAnd
-            | BinaryOp::BitOr
-            | BinaryOp::BitXor
-            | BinaryOp::ShiftLeft
-            | BinaryOp::ShiftRight => {
-                if *left == Type::Unknown || *right == Type::Unknown {
-                    let known = if *left != Type::Unknown {
-                        left
-                    } else {
-                        right
-                    };
-                    return Ok(known.clone());
-                }
-                if !left.is_integer() {
-                    return Err(Diagnostic::error(
-                        "E3012",
-                        format!(
-                            "bitwise operator requires integer type, found `{}`",
-                            left.display_name()
-                        ),
-                        *span,
-                    ));
-                }
-                if !right.is_integer() {
-                    return Err(Diagnostic::error(
-                        "E3012",
-                        format!(
-                            "bitwise operator requires integer type, found `{}`",
-                            right.display_name()
-                        ),
-                        *span,
-                    ));
-                }
-                Ok(left.clone())
-            }
-        }
     }
 }
